fixed bug on 32bit systems where size_t does not refer to a 64 bit data type.

+++ src/rastergis/RSGISCreateNewAttributeTable.cpp
@@ -196,7 +196,7 @@ namespace rsgis{namespace rastergis{
     }
     
     
-    RSGISPopAttributeTablePxlCountCalcImg::RSGISPopAttributeTablePxlCountCalcImg(int numberOutBands, RSGISAttributeTable *attTable, unsigned int pxlCountIdx):rsgis::img::RSGISCalcImageValue(numberOutBands)
+    RSGISPopAttributeTablePxlCountCalcImg::RSGISPopAttributeTablePxlCountCalcImg(int numberOutBands, RSGISAttributeTable *attTable, size_t pxlCountIdx):rsgis::img::RSGISCalcImageValue(numberOutBands)
     {
         this->attTable = attTable;
         this->pxlCountIdx = pxlCountIdx;
